+++ src/Symfony/Component/Security/Core/Event/ProgrammaticLoginEvent.php
@@ -1,0 +1,37 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Security\Core\Event;
+
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Contracts\EventDispatcher\Event;
+
+/**
+ * This event is dispatch when login programmatically.
+ *
+ * @internal
+ *
+ * @author Antoine Makdessi <amakdessi@me.com>
+ */
+class ProgrammaticLoginEvent extends Event
+{
+    private UserInterface $user;
+
+    public function __construct(UserInterface $user)
+    {
+        $this->user = $user;
+    }
+
+    public function getUser(): UserInterface
+    {
+        return $this->user;
+    }
+}